cdproject2 fix cmt commit

+++ project starter code/util/util.js
@@ -1,5 +1,6 @@
 import fs from "fs";
 import Jimp from "jimp";
+import axios from "axios";
 
 
 // filterImageFromURL
@@ -12,16 +13,16 @@ import Jimp from "jimp";
  export async function filterImageFromURL(inputURL) {
   return new Promise(async (resolve, reject) => {
     try {
-      const photo = await Jimp.read(inputURL);
-      const outpath =
-        "/tmp/filtered." + Math.floor(Math.random() * 2000) + ".jpg";
+      const photoBuf = await axios.get(inputURL, { responseType: "arraybuffer" });
+      const photoBuffer = Buffer.from(photoBuf.data);
+      const photo = await Jimp.read(photoBuffer);
+
+      const outpath =  "/tmp/filtered." + Math.floor(Math.random() * 2000) + ".jpg";
       await photo
         .resize(256, 256) // resize
         .quality(60) // set JPEG quality
         .greyscale() // set greyscale
-        .write(outpath, (img) => {
-          resolve(outpath);
-        });
+        .write(outpath, () => resolve(outpath));
     } catch (error) {
       reject(error);
     }
@@ -35,6 +36,14 @@ import Jimp from "jimp";
 //    files: Array<string> an array of absolute paths to files
  export async function deleteLocalFiles(files) {
   for (let file of files) {
-    fs.unlinkSync(file);
+    try {
+      if (fs.existsSync(file)) {
+        fs.unlinkSync(file);
+      } else {
+        console.log(`File not found: ${file}`);
+      }
+    } catch (error) {
+      console.error(`Error deleting file ${file}:`, error);
+    }
   }
 }
